// projects/h21-map/src/dto/interfaces/i-address.ts
export interface IPointAddress {
    country: string;
    city: string;
    house: string;
    postCode: string;
    countryCode: string;
    district: string;
    street: string;
    description: string;
}

export class PointAddress {
    country: string;
    city: string;
    house: string;
    postCode: string;
    countryCode: string;
    district: string;
    street: string;
    description: string;
}